Modelo/requirements.txt
absl-py==2.1.0
altair==5.5.0
annotated-types==0.7.0
anyio 
argon2-cffi 
argon2-cffi-bindings 
arrow 
asttokens 
astunparse==1.6.3
async-lru 
attrs 
Babel 
backcall 
beautifulsoup4
bleach 
blinker==1.9.0
boto3==1.35.43
botocore==1.35.43
Brotli 
cached-property 
certifi 
cffi 
charset-normalizer 
click==8.1.7
cloudpickle==3.1.1
colorama
comm 
contourpy==1.3.0
cycler 
dalex==1.7.2
debugpy 
decorator 
deeplake==3.9.26
defusedxml 
dill==0.3.9
entrypoints 
exceptiongroup 
executing 
fastjsonschema 
filelock==3.17.0
flatbuffers==24.3.25
fonttools 
fqdn 
fsspec==2025.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.1
h11 
h2 
h5py==3.12.1
highdicom==0.22.0
hpack==4.0.0
httpcore 
httpx 
humbug==0.3.2
hyperframe 
idna 
imageio==2.36.0
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata 
importlib_resources 
imutils==0.5.4
ipykernel
ipython 
ipython-genutils==0.2.0
isoduration 
jedi 
Jinja2 
jmespath==1.0.1
joblib 
json5 
jsonpointer 
jsonschema 
jsonschema-specifications 
jupyter-events 
jupyter-highlight-selected-word==0.2.0
jupyter-lsp 
jupyter_client 
jupyter_contrib_core==0.4.2
jupyter_core 
jupyter_nbextensions_configurator==0.6.4
jupyter_server 
jupyter_server_terminals 
jupyterlab==4.3.4
jupyterlab_pygments 
jupyterlab_server 
keras==2.14.0
kiwisolver 
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
MarkupSafe 
matplotlib==3.9.4
matplotlib-inline 
metrics==0.3.3
mistune 
mkl-service==2.4.0
mkl_fft 
mkl_random 
ml-dtypes==0.2.0
mpmath==1.3.0
multiprocess==0.70.17
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
narwhals==1.31.0
natsort==8.4.0
nbclient 
nbconvert 
nbformat 
nest_asyncio 
networkx==3.2.1
notebook==7.3.2
notebook_shim 
numba==0.60.0
numpy 
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
overrides 
packaging 
pandas==2.2.3
pandocfilters 
parso 
pathlib2==2.3.7.post1
pathos==0.3.3
pathspec==0.5.5
pickleshare 
pillow 
pillow-jpls==1.3.2
pkgutil_resolve_name 
platformdirs 
plotly==5.24.1
ply==3.11
pox==0.3.5
ppft==1.7.6.9
prometheus_client 
prompt_toolkit 
protobuf==4.25.5
psutil 
pure_eval 
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser 
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydicom==2.4.4
Pygments==2.18.0
PyJWT==2.9.0
pyparsing 
PyQt5==5.15.10
PyQt5-sip 
PySocks 
python-dateutil 
python-json-logger 
pytz 
pywin32==305.1
pywinpty 
PyYAML 
pyzmq 
referencing 
requests 
requests-oauthlib==2.0.0
rfc3339-validator 
rfc3986-validator 
rpds-py 
rsa==4.9
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn 
scipy 
seaborn==0.13.2
Send2Trash 
shap==0.47.2
sip 
six
slicer==0.0.8
smmap==5.0.2
sniffio 
soupsieve 
SpeechRecognition==3.14.1
stack-data
streamlit==1.43.2
streamlit-lottie==0.0.5
streamlit-shap==1.0.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado 
threadpoolctl
tifffile==2024.8.30
tinycss2 
toml==0.10.2
tomli 
torch==2.6.0
torchvision==0.21.0
tornado 
tqdm==4.66.5
traitlets 
types-python-dateutil
typing-utils 
typing_extensions
tzdata==2024.2
uib_vfeatures==0.7
ultralytics==8.3.75
ultralytics-thop==2.0.14
unicodedata2 
uri-template 
urllib3==1.26.20
watchdog==6.0.0
wcwidth 
webcolors
webencodings 
websocket-client 
Werkzeug==3.1.3
win-inet-pton 
wrapt==1.14.1
xgboost==2.1.4
zipp